refactor(shelter): useSearch hook 이름을 useSearchFilter 로 변경
